[refactor] bid 작성 스크린 재사용성 component로 분리
