+++ client/src/Navigator/MainTab/Designer/bidTabScreen.js
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { Stack } from '../../../../App';
 import { createStackNavigator } from '@react-navigation/stack';
 
+import UpdateBidScreen from '../../../Screen/02_SecondTab/Designer/updateBidScreen';
 import BidMainScreen from '../../../Screen/02_SecondTab/Designer/bidMainScreen';
 
 const bidStack = createStackNavigator();
@@ -15,6 +16,13 @@ function BidStackScreen() {
         options={{
           title: 'bid',
           headerShown: false,
+        }}
+      />
+      <bidStack.Screen
+        name="updateBid"
+        component={UpdateBidScreen}
+        options={{
+          title: 'bid',
         }}
       />
     </Stack.Navigator>
