+++ client/src/Screen/02_SecondTab/Designer/updateBidScreen.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+function UpdateBidScreen() {
+  return (
+    <View>
+      <Text>ddd</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({});
+
+export default UpdateBidScreen;
